支持 OAuth 登录

+++ server/controller/rbac_rolebind.go
@@ -3,6 +3,7 @@ package controller
 import (
 	"github.com/gin-gonic/gin"
 	"github.com/sanmuyan/xpkg/xresponse"
+	"github.com/sanmuyan/xpkg/xutil"
 	"github.com/sirupsen/logrus"
 	"wukong/pkg/util"
 	"wukong/server/model"
@@ -23,10 +24,14 @@ func GetRoleBinds(c *gin.Context) {
 func CreateRoleBinds(c *gin.Context) {
 	var roleBinds []model.RoleBind
 	if err := c.ShouldBindJSON(&roleBinds); err != nil {
-		util.Respf().Fail(xresponse.HttpBadRequest).WithError(util.NewRespError(err)).Response(util.GinRespf(c))
+		util.Respf().Fail(xresponse.HttpBadRequest).Response(util.GinRespf(c))
 		return
 	}
 	for _, roleBind := range roleBinds {
+		if xutil.IsZero(roleBind.ResourceID, roleBind.RoleID) {
+			util.Respf().Fail(xresponse.HttpBadRequest).Response(util.GinRespf(c))
+			return
+		}
 		if err := svc.CreateRoleBind(&roleBind); err != nil {
 			logrus.Errorf("创建角色绑定: %s", err.Err)
 			util.Respf().FailWithError(err).Response(util.GinRespf(c))
@@ -39,10 +44,14 @@ func CreateRoleBinds(c *gin.Context) {
 func DeleteRoleBinds(c *gin.Context) {
 	var roleBinds []model.RoleBind
 	if err := c.ShouldBindJSON(&roleBinds); err != nil {
-		util.Respf().Fail(xresponse.HttpBadRequest).WithError(util.NewRespError(err)).Response(util.GinRespf(c))
+		util.Respf().Fail(xresponse.HttpBadRequest).Response(util.GinRespf(c))
 		return
 	}
 	for _, roleBind := range roleBinds {
+		if xutil.IsZero(roleBind.ID) {
+			util.Respf().Fail(xresponse.HttpBadRequest).Response(util.GinRespf(c))
+			return
+		}
 		if err := svc.DeleteRoleBind(&roleBind); err != nil {
 			logrus.Errorf("删除角色绑定: %s", err.Err)
 			util.Respf().FailWithError(err).Response(util.GinRespf(c))
